Remove unused function fetchUsers().
The only reference to fetchUsers() was removed with 8362d617 (the
function was named sendGetUsers() at that time). fetchUsers was also the
only place where INIT_USERS got dispached, so we can remove that Action
as well.
Currently, we load an initial list of all users when initializing an
account, and capture EVENT_USER_ADD events.