user selectors: Remove unused functions.
Remove getUsersStatusOffline, getUsersStatusIdle, and
getUsersStatusActive. Apparently,
> this was created for a UI version of the thing we call now
UserPickerCard and that was not implemented exactly as intended.

Also remove getActiveUsers, since it was only used by the three
getUserStatus* functions. In fact, getActiveUsers was broken and would
always have returned an empty list, because the `is_active` property
doesn't exist anymore. `is_active` will be removed in the next commit.